+++ src/common/config/multer.ts
@@ -1,9 +1,7 @@
 import multer from 'multer';
 
 const multerStorage = multer.memoryStorage();
-
 const multerUpload = multer({
 	storage: multerStorage,
 });
-
 export { multerUpload };
