+++ src/common/interfaces/helper.ts
@@ -2,5 +2,4 @@ interface IHashData {
 	id?: string;
 	token?: string;
 }
-
 export { IHashData };
