+++ src/controllers/index.ts
@@ -1,3 +1,2 @@
 export * from './auth';
-export * from './errorController';
 export * from './auth/forgotPassword';
