+++ src/models/userModel.ts
@@ -173,7 +173,6 @@ userSchema.pre('save', async function (next) {
 			this.email,
 			this.phoneNumber,
 			this.photo,
-			//this.address.length,
 			this.gender,
 			this.isIdVerified,
 			this.isMobileVerified,
