+++ src/routes/authRouter.ts
@@ -2,7 +2,6 @@ import {
 	completeTimeBased2fa,
 	get2faCodeViaEmail,
 	forgotPassword,
-	protect,
 	resendVerification,
 	resetPassword,
 	session,
@@ -13,6 +12,7 @@ import {
 	verifyEmail,
 	verifyTimeBased2fa,
 } from '@/controllers';
+import { Protect } from '@/middlewares/protect';
 import { Router } from 'express';
 
 const router = Router();
@@ -26,7 +26,7 @@ router.post('/resend-verification', resendVerification);
 router.post('/2fa/code/email', get2faCodeViaEmail);
 router.post('/2fa/time/verify', verifyTimeBased2fa);
 
-router.use(protect); // Protect all routes after this middleware
+router.use(Protect); // Protect all routes after this middleware
 router.get('/session', session);
 router.get('/signout', signOut);
 router.post('/2fa/time/setup', setupTimeBased2fa);
