+++ src/routes/userRouter.ts
@@ -1,13 +1,13 @@
 import { multerUpload } from '@/common/config/multer';
-import { protect } from '@/controllers';
-import { editUserProfile } from '@/controllers/auth/editUserProfile';
+import { editUserProfile } from '@/controllers/user/editUserProfile';
 import { deleteAccount, restoreAccount, updateProfilePhoto } from '@/controllers/user';
 import express from 'express';
+import { Protect } from '@/middlewares/protect';
 const router = express.Router();
 
 router.post('/restore', restoreAccount);
 
-router.use(protect); // Protect all routes after this middleware
+router.use(Protect); // Protect all routes after this middleware
 router.post('/updateProfile', editUserProfile);
 router.post('/profile-photo', multerUpload.single('photo'), updateProfilePhoto);
 router.delete('/', deleteAccount);
